Make changes as recommended https://github.com/SWivid/F5-TTS/pull/363 for easier maintainabilty, modifed duration to be compatible with current repo
