+++ src/f5_tts/infer/examples/basic/basic.toml
@@ -8,4 +8,4 @@ gen_text = "I don't really care what you call me. I've been a silent spectator, 
 gen_file = ""
 remove_silence = false
 output_dir = "tests"
-predict_duration = false
+duration_model = false
